requirements.txt
alpha_vantage
pandas
numpy
sklearn
keras
tensorflow-gpu==1.15
matplotlib